Avoid 'conversion may change value' cases in krcolorcache.cpp
Avoid 'conversion may change value' warnings in krcolorcache.cpp
by removing a cast that was not needed.

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/57

+++ krusader/Panel/krcolorcache.cpp
@@ -152,7 +152,7 @@ void KrColorSettingsImpl::loadFromConfig()
     names = KrColorSettingNames::getNumNames();
     for (auto & name : names) {
         if (!group.readEntry(name, QString()).isEmpty()) {
-            m_numValues[name] = group.readEntry(name, (long long)0);
+            m_numValues[name] = group.readEntry(name, 0);
         }
     }
     names = KrColorSettingNames::getBoolNames();
